Добавлены проверки в сервисах

+++ src/wishes/wishes.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, ForbiddenException } from '@nestjs/common';
 import { CreateWishDto } from './dto/create-wish.dto';
 import { UpdateWishDto } from './dto/update-wish.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -17,7 +17,7 @@ export class WishesService {
   constructor(
     @InjectRepository(Wish)
     private wishesRepository: Repository<Wish>,
-  ) {}
+  ) { }
 
   async create(createWishDto: CreateWishDto): Promise<Wish> {
     return await this.wishesRepository.save(createWishDto);
@@ -35,14 +35,28 @@ export class WishesService {
     params: FindOptionsWhere<Wish>,
     updateWishDto: UpdateWishDto,
   ) {
+    const wish = await this.wishesRepository.findOne({ relations: { owner: true } });
+    if (wish.owner) {
+      throw new ForbiddenException('Нельзя изменять чужие подарки.');
+    }
     return this.wishesRepository.update(params, updateWishDto);
   }
 
   async removeOne(params: FindOptionsWhere<Wish>): Promise<DeleteResult> {
+    const wish = await this.wishesRepository.findOne({ relations: { owner: true } });
+    if (wish.owner) {
+      throw new ForbiddenException('Нельзя удалять чужие подарки.');
+    }
     return this.wishesRepository.delete(params);
   }
 
   async getRaise(amount: number, id: number): Promise<UpdateResult> {
+    const wish = await this.wishesRepository.findOne({where: { id: +id }});
+    if (wish.price && wish.raised > 0) {
+      throw new ForbiddenException(
+        'Вы не можете изменять стоимость подарка, если уже есть желающие скинуться.'
+      )
+    }
     const updatedWish = await this.wishesRepository
       .createQueryBuilder()
       .update(Wish)
